pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "google-generativeai>=0.8.5",
    "librosa>=0.11.0",
    "numpy>=2.2.6",
    "sift-stack-py>=0.7.0",
    "soundfile>=0.13.1",
    "werkzeug>=3.1.3",
]
